fix(test): harden script loader helper
