+++ tests/conftest.py
@@ -2,30 +2,7 @@
 
 from __future__ import annotations
 
-import importlib.util
-import sys
-from pathlib import Path
-from types import ModuleType
-
 import pytest
-
-
-def load_script_module(module_name: str, path: Path) -> ModuleType:
-    """Load a repo script by file path without putting ``scripts/`` on ``sys.path``."""
-    cached = sys.modules.get(module_name)
-    if isinstance(cached, ModuleType):
-        return cached
-    spec = importlib.util.spec_from_file_location(module_name, path)
-    if spec is None or spec.loader is None:
-        raise ImportError(f"Could not load spec for {module_name} at {path}")
-    module = importlib.util.module_from_spec(spec)
-    sys.modules[module_name] = module
-    try:
-        spec.loader.exec_module(module)
-    except Exception:
-        sys.modules.pop(module_name, None)
-        raise
-    return module
 
 
 def pytest_configure(config: pytest.Config) -> None:
