+++ tests/test_hook_memory_manifest.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 
 from pathlib import Path
 
-from conftest import load_script_module
+from tools.testing.script_imports import load_script_module
 
 REPO_ROOT = Path(__file__).resolve().parents[1]
